requirements.txt
absl-py==0.7.1
alembic==1.0.10
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
attrs==19.1.0
autobahn==19.6.1
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
crypto==1.4.1
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
EbookLib==0.17.1
entrypoints==0.3
flake8==3.7.7
Flask==1.0.3
Flask-Admin==1.5.3
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gast==0.2.2
grpcio==1.21.1
gunicorn==19.9.0
h5py==2.9.0
httplib2==0.12.3
idna==2.8
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.20
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jwcrypto==0.6.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lxml==4.3.3
Mako==1.0.11
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
mpmath==1.1.0
mysql-connector-python==8.0.16
Naked==0.1.31
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.3
oauth2client==4.1.3
pandas==0.24.2
pandas-datareader==0.7.0
pandocfilters==1.4.2
parso==0.4.0
pdfkit==0.6.1
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.3.1
pylint==2.3.1
PyMySQL==0.9.3
pyparsing==2.4.0
pyrsistent==0.15.1
python-dateutil==2.8.0
python-editor==1.0.4
python-jwt==3.2.4
pytz==2019.1
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.4
reportlab==3.5.23
requests==2.21.0
rsa==4.0
scipy==1.3.0
Send2Trash==1.5.0
shellescape==3.4.1
shiboken2==5.12.3
six==1.12.0
SQLAlchemy==1.3.4
sseclient==0.0.24
sympy==1.4
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
txaio==18.8.1
typed-ast==1.3.5
urllib3==1.24.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.11.1
WTForms==2.2.1
